pyproject.toml
# pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "bmb"
version = "1.0.0"
description = "Framework rapide pour applications web avec BMDB ORM"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "blinker==1.9.0",
    "bmdb==1.2.0",
    "click==8.3.1",
    "colorama==0.4.6",
    "Flask==3.1.2",
    "flask-cors==6.0.2",
    "greenlet==3.3.0",
    "itsdangerous==2.2.0",
    "Jinja2==3.1.6",
    "MarkupSafe==3.0.3",
    "psycopg2==2.9.11",
    "PyJWT==2.10.1",
    "python-dotenv==1.2.1",
    "PyYAML==6.0.3",
    "setuptools==80.9.0",
    "SQLAlchemy==2.0.45",
    "typing_extensions==4.15.0",
    "Werkzeug==3.1.5",
]

[project.scripts]
bmb = "bmb.cli:main"